Create keylogger
This script is a basic keylogger implemented in Python using the pynput library to capture keyboard inputs. The logged keystrokes are then color-coded based on predefined keywords. The script periodically sends the collected keyboard data to a specified server using HTTP POST requests as well as utilizing threading to schedule the sending of data at regular intervals.